Move normalized-form-fields to dataforms-layouts/ where it is used
