+++ packages/dataviews/src/dataforms-layouts/panel/dropdown.tsx
@@ -19,7 +19,7 @@ import { closeSmall } from '@wordpress/icons';
 import type { Form, FormField, NormalizedField } from '../../types';
 import { DataFormLayout } from '../data-form-layout';
 import { isCombinedField } from '../is-combined-field';
-import { DEFAULT_LAYOUT } from '../../normalize-form-fields';
+import { DEFAULT_LAYOUT } from '../normalize-form-fields';
 import SummaryButton from './summary-button';
 
 function DropdownHeader( {
